requirements.txt
numpy 
pillow 
pyyaml
opencv-python